Only show light and barrier when holding the item Fixes #1
