--- common/src/main/java/net/modgarden/barricade/mixin/BarrierBlockMixin.java
@@ -1,15 +1,0 @@
-package net.modgarden.barricade.mixin;
-
-import com.llamalad7.mixinextras.injector.ModifyReturnValue;
-import net.minecraft.world.level.block.BarrierBlock;
-import net.minecraft.world.level.block.RenderShape;
-import org.spongepowered.asm.mixin.Mixin;
-import org.spongepowered.asm.mixin.injection.At;
-
-@Mixin(value = BarrierBlock.class, priority = 500)
-public class BarrierBlockMixin {
-    @ModifyReturnValue(method = "getRenderShape", at = @At("RETURN"))
-    private RenderShape barricade$barrierBlockModel(RenderShape original) {
-        return RenderShape.MODEL;
-    }
-}
